Fix integer overflow in mp_check and mp_next for MP_MAP
An MP_MAP header encodes the number of pairs stored in the map so we
multiply the value by 2 in mp_next and mp_check. The problem is that we
don't extend the integer value before peforming multiplication so it can
lead to an integer overflow and, as a result, mp_check fails to detect
invalid MsgPack data. Fix this.

Note, we only add a test for mp_check, because adding a test for mp_next
would make it consume too much memory (> 4GB).

Needed for https://github.com/tarantool/security/issues/18