Add mp_is_loselessly_convertible_to_double
Add a new function to check if MsgPack data (whether it's integer
or floating point) can be loselessly converted to double. The
effect of the function is equivalent to this:

```
const char *p = data;
double dummy;
return mp_read_double(&p, &dummy) == 0;
```

Just a syntactic sugar if the compiler is smart enough to duplicate
the `mp_read_double` with unused stores removed (or just inline it
and eliminate unused variables).