Fix mp_compare_uint()
It wasn't correctly comparing numbers when their encoding wasn't
the most efficient. For example, if number '1' was encoded in both
arguments as 1 byte and 2 bytes, the comparator would think that
one of them is bigger.

The function was mistakenly assuming that all numbers are always
encoded as tight as possible, i.e. numbers until 0x7f are always
1 byte. Same about the other sizes.

The patch makes it more suspicious when the number types don't
match. In that case the only choice is to decode them separately
and compare. That is the slowest path now.

When the numbers have the same type, however, there is a benefit:
it is enough to keep one switch-case statement for both of them.
From that seems to be coming a significant speed up, compared to
just plain full decode of both numbers always separately.

In Tarantool's tuple.perftest on tuple_tuple_compare<FORMAT_BASIC>
case the speed doesn't seem to change. Perhaps because many
numbers had the same type and managed to hit the optimized path.

At the same time, it is clear that a plain separated decoding in
all cases is notably slower (an unconditional mp_decode_uint() for
both arguments on each mp_compare_uint()).

The results are below. The bench was run multiple times. The
operations are tuple_compare() calls. The results are medians from
several runs of each solution.

    Old incorrect comparison: 141.830 M/s
    Naive separate decoding:  137.297 M/s
    This commit's solution:   141.743 M/s

Part of tarantool/tarantool#9965